+++ lib/view_component/compiler.rb
@@ -77,6 +77,7 @@ module ViewComponent
 
         define_render_template_for
 
+        component_class.build_i18n_backend
         component_class._after_compile
 
         CompileCache.register(component_class)